+++ ansible/roles/test/files/ptftests/py3/dhcp_relay_stress_test.py
@@ -1,0 +1,39 @@
+import time
+import ptf.testutils as testutils
+from dhcp_relay_test import DHCPTest
+
+
+class DHCPContinuousStressTest(DHCPTest):
+    """
+    Keep sending packets, but don't verify form ptf side.
+    """
+    def __init__(self):
+        DHCPTest.__init__(self)
+
+    def setUp(self):
+        DHCPTest.setUp(self)
+        self.send_interval = 1 / self.test_params["pps"]
+        self.duration = self.test_params["duration"]
+        self.client_ports = self.other_client_port
+        self.client_ports.append(self.client_port_index)
+
+    def send_packet_with_interval(self, pkt, index):
+        testutils.send_packet(self, index, pkt)
+        time.sleep(self.send_interval)
+
+    def runTest(self):
+        dhcp_discover = self.create_dhcp_discover_packet(self.dest_mac_address, self.client_udp_src_port)
+        dhcp_offer = self.create_dhcp_offer_packet()
+        dhcp_request = self.create_dhcp_request_packet(self.dest_mac_address, self.client_udp_src_port)
+        dhcp_ack = self.create_dhcp_ack_packet()
+
+        start_time = time.time()
+        while time.time() - start_time <= self.duration:
+            for client_port in self.client_ports:
+                self.send_packet_with_interval(dhcp_discover, client_port)
+            for server_port in self.server_port_indices:
+                self.send_packet_with_interval(dhcp_offer, server_port)
+            for client_port in self.client_ports:
+                self.send_packet_with_interval(dhcp_request, client_port)
+            for server_port in self.server_port_indices:
+                self.send_packet_with_interval(dhcp_ack, server_port)
